backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.20
python-multipart>=0.0.6
python-telegram-bot>=20.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0